update references after refactor
