+++ components/items/weapons/PerkColumn.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from 'react'
 import { FormattedPerk } from '../../../lib/items'
 import { RadioGroup } from '@headlessui/react'
-import { Tooltip, TooltipContent, TooltipTrigger } from '../../Tooltip'
+import { Tooltip, TooltipContent, TooltipTrigger } from '../../tooltips/Tooltip'
 import Image from 'next/image'
 import { applyBungieDomain } from '../../../utils/url-handling'
-import PerkTooltip from '../../PerkTooltip'
+import PerkTooltip from '../../tooltips/PerkTooltip'
 
 type PerkColumnProps = {
   perks: FormattedPerk[]
